Adjusted unit tests
