+++ angular.json
@@ -163,6 +163,5 @@
         }
       }
     }
-  },
-  "defaultProject": "mmm-toast-demo"
+  }
 }
