+++ package.json
@@ -4,56 +4,49 @@
   "scripts": {
     "ng": "ng",
     "start": "node ./scripts/copy-styles.js demo && ng serve",
-    "build-prod": "node ./scripts/copy-styles.js && ng build mmm-toast --prod",
-    "gh-pages": "node ./scripts/copy-styles.js demo && ng build --prod --output-path docs --base-href /mmm-toast/",
-    "test-ci": "ng test mmm-toast --browsers=ChromeHeadlessCI --watch=false",
-    "test-local": "ng test mmm-toast --browsers=Chromium --watch=true",
+    "build": "node ./scripts/copy-styles.js && ng build mmm-toast --configuration production",
+    "gh-pages": "node ./scripts/copy-styles.js demo && ng build --configuration production --output-path docs --base-href /mmm-toast/",
+    "test": "ng test mmm-toast --browsers=ChromiumHeadless --watch=true",
     "lint": "ng lint mmm-toast",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.11",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/core": "~13.3.11",
-    "@angular/forms": "~13.3.11",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
-    "bootstrap": "^4.6.0",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.11.8"
+    "@angular/animations": "^14.2.8",
+    "@angular/common": "^14.2.8",
+    "@angular/compiler": "^14.2.8",
+    "@angular/core": "^14.2.8",
+    "@angular/forms": "^14.2.8",
+    "@angular/platform-browser": "^14.2.8",
+    "@angular/platform-browser-dynamic": "^14.2.8",
+    "@angular/router": "^14.2.8",
+    "bootstrap": "^5.2.2",
+    "rxjs": "^7.5.7",
+    "tslib": "^2.4.0",
+    "zone.js": "^0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular-eslint/builder": "1.1.0",
-    "@angular-eslint/eslint-plugin": "1.1.0",
-    "@angular-eslint/eslint-plugin-template": "1.1.0",
-    "@angular-eslint/schematics": "1.1.0",
-    "@angular-eslint/template-parser": "1.1.0",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~13.3.11",
-    "@types/jasmine": "^3.6.3",
-    "@types/node": "^12.19.15",
-    "@typescript-eslint/eslint-plugin": "4.3.0",
-    "@typescript-eslint/parser": "4.3.0",
-    "eslint": "^7.18.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jsdoc": "30.7.6",
-    "eslint-plugin-prefer-arrow": "1.2.2",
-    "jasmine-core": "~3.8.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.4.1",
+    "@angular-devkit/build-angular": "^14.2.7",
+    "@angular-eslint/builder": "^1.1.0",
+    "@angular-eslint/eslint-plugin": "^1.1.0",
+    "@angular-eslint/eslint-plugin-template": "^1.1.0",
+    "@angular-eslint/schematics": "^1.1.0",
+    "@angular-eslint/template-parser": "^1.1.0",
+    "@angular/cli": "^14.2.7",
+    "@angular/compiler-cli": "^14.2.8",
+    "@types/jasmine": "^3.10.6",
+    "@types/node": "^12.20.55",
+    "@typescript-eslint/eslint-plugin": "^5.41.0",
+    "@typescript-eslint/parser": "^5.41.0",
+    "eslint": "^7.32.0",
+    "fs-extra": "10.1.0",
+    "jasmine-core": "~4.3.0",
+    "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "ncp": "^2.0.0",
-    "ng-packagr": "^13.3.1",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "typescript": "^4.6.4"
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "ng-packagr": "^14.2.2",
+    "typescript": "~4.7.2"
   }
 }
