+++ projects/mmm-toast/src/lib/models/index.ts
@@ -1,0 +1,3 @@
+export * from './global-config.model';
+export * from './toast.model';
+export * from './types';
