+++ projects/mmm-toast/src/lib/models/toast.model.ts
@@ -1,0 +1,6 @@
+import { GlobalConfigModel, GlobalType } from '.';
+
+export interface ToastModel extends GlobalConfigModel {
+  type: GlobalType;
+  message: string;
+}
