+++ scripts/copy-styles.js
@@ -1,6 +1,4 @@
-const ncp = require('ncp').ncp;
-
-ncp.limit = 1;
+const fs = require('fs-extra');
 
 let destinationPath = 'dist/mmm-toast/lib/styles';
 
@@ -8,7 +6,7 @@ if (process.argv[2] === 'demo') {
   destinationPath = 'src/assets'
 }
 
-ncp('projects/mmm-toast/src/lib/styles', destinationPath, (err) => {
+fs.copy('projects/mmm-toast/src/lib/styles', destinationPath, (err) => {
   if (err) return console.log(err);
   console.log(`******************* styles successfully copied to ${destinationPath} folder *******************`);
 })
